Add files via upload
There are 2 codes. the first is a simple  code that has 3 lines and can generate a simple qr code. But the secode code can generate qr code a text or url in the way you need it you can change the color or the size of the code or the error correction. The second is advanced way to generate qr codes